backend/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
firebase-admin>=5.0.0
requests>=2.26.0
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2 